check deps on install

+++ src/dfctl/cli/cmd/install.py
@@ -33,34 +33,42 @@ class CMD(SubParser):
                         f"could not install branch '{str(group)}'\nanother branch from the same group already installed '{str(alreadyones[0])}'\nfor branch overwriting, use the force argument (-f)"
                     )
 
-                path: Path = group.path
-                syms = group.get_syms()
+                if group.check_deps()[0]:
+                    print(group.check_deps()[0])
+                    path: Path = group.path
+                    syms = group.get_syms()
 
-                installed: list = []
-                run_status: str = ""
-                for original, sym in syms.items():
-                    sym = Path(sym).expanduser()
-                    original = path / original
-                    run_status = "WRITE"
+                    installed: list = []
+                    run_status: str = ""
+                    for original, sym in syms.items():
+                        sym = Path(sym).expanduser()
+                        original = path / original
+                        run_status = "WRITE"
 
-                    while True:
-                        try:
-                            sym.symlink_to(original)
-                            break
-                        except FileExistsError:
-                            if sym.is_symlink():
-                                if sym.readlink() == original:
-                                    run_status = "ALREADY"
-                                    break
-                                else:
-                                    sym.unlink() if sym.is_file() else rmtree(sym)
-                                    run_status = "OVERWRITE"
-                    installed.append((run_status, sym, original, _styles[run_status]))
+                        while True:
+                            try:
+                                sym.symlink_to(original)
+                                break
+                            except FileExistsError:
+                                if sym.is_symlink():
+                                    if sym.readlink() == original:
+                                        run_status = "ALREADY"
+                                        break
+                                    else:
+                                        sym.unlink() if sym.is_file() else rmtree(sym)
+                                        run_status = "OVERWRITE"
+                        installed.append(
+                            (run_status, sym, original, _styles[run_status])
+                        )
 
-                console.log(f"[bold green]Installed[/] [bold blue]'{str(group)}'")
-                for x in installed:
-                    console.log(f"\t{x[3]}{f'{x[0]}[/]':<9} ({beautypath(x[1])} > {
-                        beautypath(x[2])})")
+                    console.log(f"[bold green]Installed[/] [bold blue]'{str(group)}'")
+                    for x in installed:
+                        console.log(f"\t{x[3]}{f'{x[0]}[/]':<9} ({beautypath(x[1])} > {
+                            beautypath(x[2])})")
+                else:
+                    console.log(
+                        f"[bold red] deps doesn't line up for group '{str(group)}'"
+                    )
 
     def setup(self, subparser):
         subparser.add_argument("target", help="group target")
